% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{print.event_log_schema}
\alias{print.event_log_schema}
\title{Print an event_log_schema object}
\usage{
\method{print}{event_log_schema}(x, ...)
}
\arguments{
\item{x}{An `event_log_schema` object.}

\item{...}{Ignored.}
}
\value{
`x`, invisibly.
}
\description{
Print an event_log_schema object
}
