% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventviz-package.R
\docType{package}
\name{eventviz-package}
\alias{eventviz}
\alias{eventviz-package}
\title{eventviz: Event Log Visualisation}
\description{
A general-purpose package for visualising timestamped event logs as
location-band timelines, staircase stage diagrams, cohort facets, and
aggregate/statistical summaries. See [plot_patient_journey()] to get
started.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jaspercain01.github.io/event-driven-visualisation/}
}

}
\author{
\strong{Maintainer}: Jasper Cain \email{jcainsecondemail@gmail.com}

}
\keyword{internal}
